api/requirements.txt
# financial libraries
finnhub-python
yfinance
sec_api

# ai libraries
openai

# others libraries
python-dotenv
requests
pandas

